refactor: move usercontext and venuecontext into own context folder in components, update imports
