+++ src/main.jsx
@@ -1,8 +1,8 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
-import { UserProvider } from "./components/UserContext.jsx";
-import { VenueProvider } from "./components/VenueContext.jsx";
+import { UserProvider } from "./components/context/UserContext.jsx";
+import { VenueProvider } from "./components/context/VenueContext.jsx";
 import "./index.css";
 import App from "./App.jsx";
